lib: libc: common: provide a common clock() implementation
The `clock()` function was originally part of ISO C89, and it is
also a POSIX extension of the C standard.

Both newlib and picolibc have implementations of `clock()` that
reference the POSIX `times()` function.  However, POSIX is
optional in Zephyr while C is non-optional. Due to this, the
newlib and picolibc implementations cause a layering violation.

Provide a simple implementation here that is independent of
POSIX to avoid the layering violation.

At some future date, we should be able to require all platforms
to provide a `chosen` RTC (whether hardware-based or emulated),
which should be capable of providing better resolution
(the RTC API is capable of up to nanosecond granularity).

```devicetree
chosen {
    zephyr,rtc = &rtc0;
};
```

Signed-off-by: Chris Friedt <cfriedt@meta.com>